requirements.txt
argparse
img2pdf
qrcode
dateutil
Pillow
